+++ CHANGELOG.md
@@ -7,6 +7,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Bugfix
 ### Client
 
+# 2.1.2 [13/12/2022]
+- Version (1.0.0 RC3)
+###Fixed
+- Navigation bar color displaying the wrong color in the Loading Fragment
+- Make Order Success Page scrollable to accommodate smaller screen devices
+- Terms of service bottom sheet dialog not scrolling properly
+
+###Changed
+- Remove 'Download Receipt Button'
+- Update price text to display correctly
+- Change the drawable in the register form for the Location field
+- Secure API keys inside api.properties file
+- Update .gitignore file
+
+
+
 # 2.1.1 [09/12/2022]
 - Version (1.0.0 RC2)
 ###Fixed
@@ -15,9 +31,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ###Changed
 - In checkout page sales agents list is now dynamic
-- Added M-PESA SDK
-- Added history item click details
-- Removed unused imports and classes
+- Add M-PESA SDK
+- Add history item click details
+- Remove unused imports and classes
 
 
 # 2.1.0 [06/12/2022]